omit stddef.h

+++ src/memory/TNew.hpp
@@ -4,8 +4,8 @@
 	#ifndef TKLB_NO_STDLIB
 		#include <new>
 	#else // TKLB_NO_STDLIB
-		#include <stddef.h>
-		inline void* operator new(size_t, void* __p) { return __p; }
+		#include "../types/TTypes.hpp"
+		inline void* operator new(unsigned long, void* __p) { return __p; }
 	#endif // TKLB_NO_STDLIB
 
 #endif // _TKLB_NEW
